Create script.js
add code to make extension work.

This code is for a google chrome extension. It will make a request to our external server API, which will send back a list of brands. The extension will then look for brands on the webpage that it is on, and compare this to the list sent back from the API. Any matching brands will then have some DOM manipulation done to them on this webpage.